+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.0.0 
 - Class representation (`__repr__`) added on NepaliTimeZone
+- Refactored converter
 
 ## v0.5.6 - (July 7, 2022)
 - Test for automatic package publish